backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-socketio[asgi]>=5.11
pydantic-settings>=2.2
sqlalchemy>=2.0
aiosqlite>=0.20
asyncpg>=0.29
alembic>=1.13
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
